refactor: move redirection role request ti main screen
